Cleaned up code and added xml comments
