+++ src/MefContrib.Hosting.Conventions.Tests/PartRegistryTests.cs
@@ -1,11 +1,7 @@
 namespace MefContrib.Hosting.Conventions.Tests
 {
     using System;
-    using System.Collections.Generic;
-    using System.IO;
     using System.Linq;
-    using System.Reflection;
-
     using MefContrib.Hosting.Conventions.Configuration;
     using NUnit.Framework;
 
@@ -81,100 +77,6 @@ namespace MefContrib.Hosting.Conventions.Tests
                 registry.TypeScanner.GetTypes(x => true);
 
             results.Count().ShouldEqual(2);
-        }
-    }
-
-    public class AssemblyFactory : IDisposable
-    {
-        public AssemblyFactory()
-        {
-            this.Assemblies = new List<AssemblyInfo>();
-            this.AssemblyDirectory = CreateTemporaryDirectory();
-        }
-
-        public AssemblyInfo Build(string code)
-        {
-            var builtAssembly =
-                CSharpAssemblyFactory.Compile(code, this.GenerateTemporaryAssemblyName());
-
-            this.Assemblies.Add(new AssemblyInfo(builtAssembly));
-
-            return this.Assemblies.Last();
-        }
-
-        public IList<AssemblyInfo> Assemblies { get; private set; }
-
-        public DirectoryInfo AssemblyDirectory { get; set; }
-
-        private static DirectoryInfo CreateTemporaryDirectory()
-        {
-            var tempDirectoryPath =
-               Path.Combine(Path.GetTempPath(),
-               Path.GetFileNameWithoutExtension(Path.GetRandomFileName()));
-
-            return Directory.CreateDirectory(tempDirectoryPath);
-        }
-
-        private string GenerateTemporaryAssemblyName()
-        {
-            var assemblyName =
-                Path.Combine(this.AssemblyDirectory.FullName,
-                string.Concat(Path.GetFileNameWithoutExtension(Path.GetRandomFileName()), ".dll"));
-
-            return assemblyName;
-        }
-
-        public void Dispose()
-        {
-            foreach (var assemblyInfo in this.Assemblies)
-            {
-                assemblyInfo.Dispose();
-            }
-
-            try
-            {
-                Directory.Delete(this.AssemblyDirectory.FullName, true);
-            }
-            catch
-            {
-            }
-        }
-    }
-
-    public class AssemblyInfo : IDisposable
-    {
-        public AssemblyInfo(Assembly assembly)
-        {
-            if (assembly == null)
-            {
-                throw new ArgumentNullException("assembly", "The assembly cannot be null");
-            }
-
-            this.Assembly = assembly;
-        }
-
-        public Assembly Assembly { get; private set; }
-
-        public void Dispose()
-        {
-            try
-            {
-                File.Delete(this.Assembly.FullName);
-            }
-            catch
-            {
-            }
-            
-        }
-
-        public static implicit operator Assembly(AssemblyInfo info)
-        {
-            return info.Assembly;
-        }
-
-        public static implicit operator string(AssemblyInfo info)
-        {
-            return info.Assembly.Location;
         }
     }
 }
